+++ Cargo.toml
@@ -200,6 +200,7 @@ members = [
     "core/sgx",
     "core/version",
     "core/vpn",
+    "exe-unit/components/counters",
     "exe-unit/components/gsb-http-proxy",
     "exe-unit",
     "exe-unit/runtime-api",
@@ -224,7 +225,6 @@ members = [
     "test-utils/test-framework",
     "test-utils/test-framework/framework-macro",
     "test-utils/test-framework/framework-basic",
-    "exe-unit/components/gsb-http-proxy"
 ]
 
 [workspace.dependencies]
@@ -309,6 +309,7 @@ ya-utils-process = { path = "utils/process" }
 ya-diesel-utils = { path = "utils/diesel-utils" }
 ya-metrics = { path = "core/metrics" }
 ya-provider = { path = "agent/provider" }
+ya-counters = { path = "exe-unit/components/counters" }
 ya-gsb-http-proxy = { path = "exe-unit/components/gsb-http-proxy" }
 
 ## TEST UTILS
