+++ exe-unit/Cargo.toml
@@ -56,6 +56,7 @@ ya-utils-networking = { version = "0.2", default-features = false, features = [
 ] }
 ya-packet-trace = { git = "https://github.com/golemfactory/ya-packet-trace" }
 ya-gsb-http-proxy = { path = "../exe-unit/components/gsb-http-proxy" }
+ya-counters = { path = "../exe-unit/components/counters" }
 
 actix = { version = "0.13", default-features = false }
 actix-rt = "2.7"
