+++ exe-unit/components/counters/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "ya-counters"
+version = "0.1.0"
+edition = "2021"
+description = "Runtime counters generating data for Provider's metrics"
+
+[dependencies]
+ya-agreement-utils = { workspace = true }
+
+actix = "0.13"
+actix-rt = "2.7"
+chrono = "0.4"
+log = "0.4"
+serde = { version = "^1.0", features = ["derive"] }
+serde_json = "1.0"
+serde_yaml = "0.9"
+thiserror = "1.0"
+tokio = { version = "1", features = [
+    "process",
+    "signal",
+    "time",
+    "net",
+    "rt-multi-thread",
+] }
