+++ exe-unit/components/counters/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod error;
+pub mod counters;
+pub mod service;
+pub mod message;
+
+pub type Result<T> = std::result::Result<T, error::MetricError>;
