+++ exe-unit/src/bin.rs
@@ -14,7 +14,7 @@ use ya_exe_unit::logger::*;
 use ya_exe_unit::manifest::ManifestContext;
 use ya_exe_unit::message::{GetState, GetStateResponse, Register};
 use ya_exe_unit::runtime::process::RuntimeProcess;
-use ya_exe_unit::service::metrics::MetricsService;
+use ya_exe_unit::service::counters;
 use ya_exe_unit::service::signal::SignalMonitor;
 use ya_exe_unit::state::Supervision;
 use ya_exe_unit::{ExeUnit, ExeUnitContext};
@@ -323,7 +323,7 @@ async fn run() -> anyhow::Result<()> {
 
     let (tx, rx) = oneshot::channel();
 
-    let metrics = MetricsService::try_new(&ctx, Some(10000), ctx.supervise.hardware)?.start();
+    let metrics = counters::try_new(&ctx, Some(10000), ctx.supervise.hardware)?.start();
     let transfers = TransferService::new((&ctx).into()).start();
     let runtime = RuntimeProcess::new(&ctx, cli.binary).start();
     let exe_unit = ExeUnit::new(tx, ctx, metrics, transfers, runtime).start();
